Load pipe list json via http
Replaced hard-coded json string with a real (though currently static) json doc
fetched via HTTP.

The image and text views in the pipe_item view now start invisible and are made
visible only after the pipe image was loaded from the server. Otherwise the
text views are badly positioned, due to the relative layout. In the future there
could be an animated wheel for each image that is to be loaded.